refactor(native): standardize platform source names
