+++ proton/internal/native/ffi/src/engine/cef_linux/linux_client.c
@@ -1,6 +1,6 @@
 #if defined(__linux__)
 
-#include "internal.h"
+#include "linux_internal.h"
 #include "../../proton_config.h"
 #include "../../proton_event.h"
 #include "../../proton_json.h"
