+++ proton/internal/native/ffi/src/engine/cef_linux/linux_menu.c
@@ -1,6 +1,6 @@
 #if defined(__linux__)
 
-#include "internal.h"
+#include "linux_internal.h"
 
 #include "../../proton_event.h"
 #include "../cef_common/message.h"
