+++ proton/internal/native/ffi/src/engine/cef_linux/linux_titlebar.c
@@ -1,6 +1,6 @@
 #if defined(__linux__)
 
-#include "internal.h"
+#include "linux_internal.h"
 
 #include <stdint.h>
 #include <string.h>
