+++ proton/internal/native/ffi/src/engine/cef_linux/linux_view.c
@@ -1,6 +1,6 @@
-#if defined(_WIN32)
+#if defined(__linux__)
 
-#include "internal.h"
+#include "linux_internal.h"
 
 #include "../../proton_config.h"
 #include "../../proton_event.h"
@@ -10,10 +10,10 @@
 #include "../cef_common/strings.h"
 #include "../cef_common/view_events.h"
 
-#define PROTON_ENGINE_REF_INCREMENT(refs) InterlockedIncrement(&(refs)->refs)
-#define PROTON_ENGINE_REF_DECREMENT(refs) InterlockedDecrement(&(refs)->refs)
-#define PROTON_ENGINE_REF_LOAD(refs) InterlockedCompareExchange(&(refs)->refs, 0, 0)
-#define PROTON_ENGINE_REF_STORE(refs, value) InterlockedExchange(&(refs)->refs, value)
+#define PROTON_ENGINE_REF_INCREMENT(refs) atomic_fetch_add_explicit(&(refs)->refs, 1, memory_order_relaxed)
+#define PROTON_ENGINE_REF_DECREMENT(refs) (atomic_fetch_sub_explicit(&(refs)->refs, 1, memory_order_acq_rel) - 1)
+#define PROTON_ENGINE_REF_LOAD(refs) atomic_load_explicit(&(refs)->refs, memory_order_acquire)
+#define PROTON_ENGINE_REF_STORE(refs, value) atomic_store(&(refs)->refs, value)
 #include "../cef_common/ref_count.h"
 #undef PROTON_ENGINE_REF_INCREMENT
 #undef PROTON_ENGINE_REF_DECREMENT
@@ -25,8 +25,8 @@
 #include "include/capi/cef_values_capi.h"
 #include "include/internal/cef_string.h"
 
-#define WIN32_LEAN_AND_MEAN
-#include <windows.h>
+#include <gdk/gdkx.h>
+#include <X11/Xlib.h>
 
 #include <math.h>
 #include <stdint.h>
@@ -35,14 +35,14 @@
 
 // MARK: - Web contents views
 //
-// A view is an extra child browser hosted inside a window's client area,
+// A view is an extra child browser hosted inside a window's browser host,
 // following the Electron WebContentsView model: explicit top-left bounds,
 // visibility, z-order, and an independent load target. Struct lifetime is
-// owned by the window: views are only freed from proton_engine_window_free
-// once every view has finalized, so native ABI view slots stay valid for the
-// whole window lifetime. Close semantics mirror the macOS engine: do_close
-// takes over from CEF's default (which would post WM_CLOSE to the frame
-// window) and tears down the browser's child HWND instead.
+// owned by the window: views are only freed from the window's storage
+// teardown once every view has finalized, so native ABI view slots stay
+// valid for the whole window lifetime. Close semantics mirror the macOS
+// engine: do_close takes over from CEF's default (which would post a delete
+// event to the frame window) and destroys the browser's X window instead.
 
 static proton_engine_display_handler_t g_display_handler;
 static uint64_t g_next_view_native_id = 1;
@@ -62,15 +62,7 @@ void proton_engine_window_free_views(proton_engine_window_t *window) {
     proton_engine_view_t *next = view->next;
     proton_browser_session_destroy(view->browser_session);
     proton_view_events_destroy(view->events);
-    if (view->client != NULL) {
-      // Drop only the engine's reference; CEF's final release after
-      // OnBeforeClose frees the client (guaranteed by the finalize gate).
-      cef_base_ref_counted_t *client_base =
-          (cef_base_ref_counted_t *)view->client;
-      view->client->view = NULL;
-      view->client = NULL;
-      client_base->release(client_base);
-    }
+    free(view->client);
     free(view);
     view = next;
   }
@@ -87,7 +79,7 @@ void proton_engine_view_finalize_if_ready(proton_engine_view_t *view) {
   if (view->client != NULL) {
     view->client->view = NULL;
   }
-  view->hwnd = NULL;
+  view->xwindow = 0;
   view->finalized = 1;
   // The window's own finalize is gated on every view being finalized; this
   // call is a no-op unless the window is waiting on exactly this view.
@@ -95,8 +87,8 @@ void proton_engine_view_finalize_if_ready(proton_engine_view_t *view) {
 }
 
 void proton_engine_window_finalize_if_ready(proton_engine_window_t *window) {
-  if (window == NULL || window->finalize_queued ||
-      !window->destroy_requested || window->browser != NULL) {
+  if (window == NULL || !window->destroy_requested ||
+      window->browser != NULL) {
     return;
   }
   for (proton_engine_view_t *view = window->views; view != NULL;
@@ -105,9 +97,6 @@ void proton_engine_window_finalize_if_ready(proton_engine_window_t *window) {
       return;
     }
   }
-  // OnBeforeClose is CEF's final browser callback, but CEF still owns and
-  // releases callback objects while unwinding it and during cef_shutdown.
-  // Keep the host storage alive until that shutdown has completed.
   proton_engine_window_defer_free(window);
 }
 
@@ -142,7 +131,7 @@ void proton_engine_window_close_views(proton_engine_window_t *window) {
   }
 }
 
-// Re-stacks view browser windows above the window's main browser view by
+// Re-stacks view browser X windows above the window's main browser by
 // ascending (z_order, native_id).
 void proton_engine_window_layout_views(proton_engine_window_t *window) {
   if (window == NULL) {
@@ -151,7 +140,7 @@ void proton_engine_window_layout_views(proton_engine_window_t *window) {
   size_t count = 0;
   for (proton_engine_view_t *view = window->views; view != NULL;
        view = view->next) {
-    if (view->hwnd != NULL && !view->closed) {
+    if (view->xwindow != 0 && view->display != NULL && !view->closed) {
       count++;
     }
   }
@@ -166,7 +155,7 @@ void proton_engine_window_layout_views(proton_engine_window_t *window) {
   size_t index = 0;
   for (proton_engine_view_t *view = window->views; view != NULL;
        view = view->next) {
-    if (view->hwnd != NULL && !view->closed) {
+    if (view->xwindow != 0 && view->display != NULL && !view->closed) {
       order[index++] = view;
     }
   }
@@ -183,8 +172,7 @@ void proton_engine_window_layout_views(proton_engine_window_t *window) {
     order[j] = current;
   }
   for (size_t i = 0; i < count; i++) {
-    SetWindowPos(order[i]->hwnd, HWND_TOP, 0, 0, 0, 0,
-                 SWP_NOMOVE | SWP_NOSIZE | SWP_NOACTIVATE);
+    XRaiseWindow(order[i]->display, order[i]->xwindow);
   }
   free(order);
 }
@@ -193,20 +181,19 @@ int CEF_CALLBACK proton_engine_do_close(
     cef_life_span_handler_t *self,
     cef_browser_t *browser) {
   (void)self;
-  proton_engine_view_t *view =
-      proton_engine_find_view_by_browser_id(proton_engine_browser_id(browser));
+  proton_engine_view_t *view = proton_engine_view_from_browser(browser);
   if (view == NULL) {
-    // Frame windows keep CEF's default close behavior (WM_CLOSE to the frame).
+    // Frame windows keep CEF's default close behavior (delete event on the
+    // top-level window).
     return 0;
   }
-  // A view browser owns no top-level window; CEF's default would post
-  // WM_CLOSE to the frame window and cancel the view close. Take over and
-  // destroy the browser's child window on the next frame-window message,
-  // which completes the teardown via WindowDestroyed without re-entering CEF.
-  if (view->hwnd != NULL) {
-    PostMessageW(view->window->hwnd, PROTON_ENGINE_WM_DESTROY_CHILD, 0,
-                 (LPARAM)view->hwnd);
-    view->hwnd = NULL;
+  // A view browser owns no top-level window; CEF's default would deliver a
+  // delete event to the frame window and cancel the view close. Take over
+  // and destroy the browser's X window, which completes the teardown via
+  // WindowDestroyed.
+  if (view->xwindow != 0 && view->display != NULL) {
+    XDestroyWindow(view->display, view->xwindow);
+    view->xwindow = 0;
     return 1;
   }
   // Windowless (headless) rendering has no child window; returning false lets
@@ -219,15 +206,14 @@ void CEF_CALLBACK proton_engine_on_title_change(
     cef_browser_t *browser,
     const cef_string_t *title) {
   (void)self;
-  proton_engine_view_t *view =
-      proton_engine_find_view_by_browser_id(proton_engine_browser_id(browser));
+  proton_engine_view_t *view = proton_engine_view_from_browser(browser);
   if (view == NULL) {
     return;
   }
   char *title_utf8 = proton_engine_cef_string_to_utf8(title);
   proton_view_events_title_updated(view->events, title_utf8);
   free(title_utf8);
-  proton_engine_signal_wait_source(view->window->runtime, PROTON_WAIT_EVENT);
+  proton_engine_signal_wait_source(PROTON_WAIT_EVENT);
 }
 
 static cef_display_handler_t *CEF_CALLBACK
@@ -254,7 +240,6 @@ static proton_engine_client_t *proton_engine_view_client_create(
   }
   proton_engine_init_ref_counted((cef_base_ref_counted_t *)&client->client.base,
                                  sizeof(client->client), &client->refs);
-  client->client.base.release = proton_engine_client_release;
   client->view = view;
   // Views wire the life span, load, display, and render handlers: life span
   // drives the close state machine, load/display feed the view event stream,
@@ -280,13 +265,21 @@ static int32_t proton_engine_view_create_browser(
   memset(&browser_settings, 0, sizeof(browser_settings));
   window_info.size = sizeof(window_info);
   browser_settings.size = sizeof(browser_settings);
+  if (!window->headless &&
+      (window->browser_host == NULL ||
+       gtk_widget_get_window(window->browser_host) == NULL)) {
+    proton_engine_set_message(error, error_len,
+                              "window is not ready for view browser creation");
+    return PROTON_ERR_INVALID_ARGUMENT;
+  }
   if (window->headless) {
     window_info.windowless_rendering_enabled = 1;
     window_info.runtime_style = CEF_RUNTIME_STYLE_ALLOY;
   } else {
-    window_info.parent_window = window->hwnd;
-    window_info.style =
-        WS_CHILD | WS_VISIBLE | WS_CLIPCHILDREN | WS_CLIPSIBLINGS;
+    GdkWindow *host_gdk_window = gtk_widget_get_window(window->browser_host);
+    view->display = GDK_WINDOW_XDISPLAY(host_gdk_window);
+    window_info.parent_window =
+        (cef_window_handle_t)GDK_WINDOW_XID(host_gdk_window);
   }
   window_info.bounds.x = view->x;
   window_info.bounds.y = view->y;
@@ -314,9 +307,9 @@ static int32_t proton_engine_view_create_browser(
         host->was_hidden(host, 1);
       }
     } else {
-      view->hwnd = host->get_window_handle(host);
-      if (!view->visible && view->hwnd != NULL) {
-        ShowWindow(view->hwnd, SW_HIDE);
+      view->xwindow = host->get_window_handle(host);
+      if (!view->visible && view->xwindow != 0 && view->display != NULL) {
+        XUnmapWindow(view->display, view->xwindow);
       }
     }
     host->base.release((cef_base_ref_counted_t *)host);
@@ -411,7 +404,7 @@ int32_t proton_engine_view_create(
     proton_engine_view_finalize_if_ready(view);
     return status;
   }
-  proton_engine_signal_wait_source(window->runtime, PROTON_WAIT_PLATFORM);
+  proton_engine_signal_wait_source(PROTON_WAIT_PLATFORM);
   *out_view = view;
   return PROTON_OK;
 }
@@ -474,12 +467,10 @@ int32_t proton_engine_view_set_bounds(proton_engine_view_t *view,
         host->base.release((cef_base_ref_counted_t *)host);
       }
     }
-  } else if (view->hwnd != NULL) {
-    SetWindowPos(view->hwnd, NULL, x, y, width, height,
-                 SWP_NOZORDER | SWP_NOACTIVATE);
+  } else if (view->xwindow != 0 && view->display != NULL) {
+    XMoveResizeWindow(view->display, view->xwindow, x, y, width, height);
   }
-  proton_engine_signal_wait_source(view->window->runtime,
-                                   PROTON_WAIT_PLATFORM);
+  proton_engine_signal_wait_source(PROTON_WAIT_PLATFORM);
   return PROTON_OK;
 }
 
@@ -500,11 +491,14 @@ int32_t proton_engine_view_set_visible(proton_engine_view_t *view,
         host->base.release((cef_base_ref_counted_t *)host);
       }
     }
-  } else if (view->hwnd != NULL) {
-    ShowWindow(view->hwnd, view->visible ? SW_SHOW : SW_HIDE);
+  } else if (view->xwindow != 0 && view->display != NULL) {
+    if (view->visible) {
+      XMapWindow(view->display, view->xwindow);
+    } else {
+      XUnmapWindow(view->display, view->xwindow);
+    }
   }
-  proton_engine_signal_wait_source(view->window->runtime,
-                                   PROTON_WAIT_PLATFORM);
+  proton_engine_signal_wait_source(PROTON_WAIT_PLATFORM);
   return PROTON_OK;
 }
 
@@ -518,8 +512,7 @@ int32_t proton_engine_view_set_z_order(proton_engine_view_t *view,
   }
   view->z_order = z_order;
   proton_engine_window_layout_views(view->window);
-  proton_engine_signal_wait_source(view->window->runtime,
-                                   PROTON_WAIT_PLATFORM);
+  proton_engine_signal_wait_source(PROTON_WAIT_PLATFORM);
   return PROTON_OK;
 }
 
@@ -545,8 +538,7 @@ int32_t proton_engine_view_load_url(proton_engine_view_t *view,
   frame->load_url(frame, &cef_url);
   cef_string_clear(&cef_url);
   frame->base.release((cef_base_ref_counted_t *)frame);
-  proton_engine_signal_wait_source(view->window->runtime,
-                                   PROTON_WAIT_PLATFORM);
+  proton_engine_signal_wait_source(PROTON_WAIT_PLATFORM);
   return PROTON_OK;
 }
 
@@ -571,8 +563,7 @@ int32_t proton_engine_view_eval(proton_engine_view_t *view,
   cef_string_clear(&code);
   cef_string_clear(&script_url);
   frame->base.release((cef_base_ref_counted_t *)frame);
-  proton_engine_signal_wait_source(view->window->runtime,
-                                   PROTON_WAIT_PLATFORM);
+  proton_engine_signal_wait_source(PROTON_WAIT_PLATFORM);
   return PROTON_OK;
 }
 
@@ -589,6 +580,5 @@ int32_t proton_engine_view_browser_command_json(proton_engine_view_t *view,
                                              view->browser, command_json,
                                              error, error_len);
 }
-
 
 #endif
