+++ proton/internal/native/ffi/src/engine/cef_mac/mac_launch_input.m
@@ -1,6 +1,6 @@
 #if defined(__APPLE__)
 
-#include "launch_input.h"
+#include "mac_launch_input.h"
 #include "../../proton_event.h"
 
 #import <AppKit/AppKit.h>
